refactor(featureditem): move add, update, remove logic to admin featureditem controller/services
