+++ backend/src/services/content/featureditem_services.js
@@ -9,24 +9,3 @@ export async function getFeaturedItemsService(section) {
     .sort({ priority: -1 })
     .populate("item");
 }
-export async function addFeaturedItemService(featuredItem) {
-  const { itemType, item, priority, starDate, endDate } = featuredItem;
-  return await Featured.create({
-    itemType,
-    item,
-    priority,
-    starDate,
-    endDate,
-  });
-}
-export async function updateFeaturedItemService(featuredItem) {
-  const { id, itemType, item, priority, starDate, endDate } = featuredItem;
-  return await Featured.findByIdAndUpdate(
-    id,
-    { itemType, item, priority, starDate, endDate },
-    { new: true },
-  );
-}
-export async function removeFeaturedItemService(id) {
-  return await Featured.findByIdAndDelete(id);
-}
